Recording and replaying all the different clock types. All clock types are written to a single file.
We record all the possible clock types by first writing the event type to the file. Then we follow it
by the actual clock value. It is possible that a different clock type is read at the same moment during replay.
We stop replay at that point and terminate the guest. In essence, the replay is fairly dumb and we assume that the loaded guest
has no idea of time.