Record tsc clock values.
QEMU by default does not do a VMEXIT on RDTSC execution. To make sure that happens, a change in KVM is necessary.
Once this change is done, QEMU will report that it doesn't understand the VMEXIT. We need to add a new VMEXIT reason
called KVM_EXIT_RDTSC. The handler for this VMEXIT retrieves values of registers and composes the tsc value.