Remove unnecessary code, where replay happens before the pre-interrupt handler function runs
We need to handle replay of interrupts differently. The TCG translation code will now
do a lookup of the IntelPT packet and then decide whether to take an interrupt or not.
The interrupt handling will happen in the 'TCG' way, meaning TCG generates the code to actually
take the interrupt.