Identify interrupt number using the address in the TIP packet + Disable 'IS_INT' flag if userspace raises an interrupt.
TIP packet contains the address to which execution, if it gets an interrupt, is supposed to jump to. Given that this
is the address of the interrupt handler function, we can use this information to compute the interrupt number. We pass this
information down when we are trying to "inject" an interrupt.

As part of this commit, we are also disabling the IS_INT flag when userspace (privilege level = 3) raises an interrupt.
Interrupts in tcg are being simulated using the "raise_interrupt" function and our replay mechanism uses it. However, this
sets the IS_INT flag always to 1, since it assumes that interrupts will only be raised if tcg finds an INT n instruction.
During replay, we don't care about this.