Maintain the fully qualified TIP address of the last batch of replayed execution.
This is because we use the previous fully qualified TIP address to evaluate and construct
the current TIP address. Some TIP addresses are partially emitted by the processor to save buffer
and must be converted into the 'full' address. Since execution happens in a batch now, we need to
maintain the last fully qualified TIP address of the previous batch.