Replay disk I/O requests.
Note that this doesn't really use the blkreplay driver.  Disk I/O events are opportunistically serviced
before the virtio disk interrupt (intno: 81) and when the virtqueue_get_buf function is called.
This change also includes fixes to make virtio disk I/O events synchronous in nature. This means getting rid of thread pools,
which is used by default. The current limitation of this approach is that it will only work if there
is a single virtqueue for virtio-blk.