Replay disk events by replaying the coroutines that complete disk I/O.
This is a change from the initial disk I/O replay mechanism we had.
All disk events in replay mode will now be stored into an event queue.
When the interrupt happens, we will compare the disk event ids with the
event ids stored in the record file, and replay the coroutine
assocated with the "event" structure whenever there is a match.