Replay disk I/O events manually.
This brings back disk I/O replay to one of the original versions. We have experimentally determined that
the guest code execution thread in TCG, needs to be manually stopped and use that stoppage to replay
disk I/O events. This means we do something similar to what we do for tx packets for network I/O. We still
need to record the completion IDs though.