Multi-core replay: Special case handling for interrupt number 251.
Interrupt number 251 is one of 3 different kinds of interprocessor interrupts
in Linux, called CALL_FUNCTION_VECTOR. We detect the occurence of int 251 in destination
CPUs by writes to MSR 2096 in the source CPU. When this happens, we stop our usual
scheduling approach to deal with int 251 in the destination CPUs. We begin
our usual scheduling approach once the destination CPUs have finished handling the
interrupt (detected via 'iretq').