Multi-core replay: Compute the current TSC counter value when IPIs are sent.
Quite frequently, we see that by the time we record the sequence of IPIs and a TSC read event,
the order in which they occur is actually different from the order in which they are recorded
(the order gets flipped, due to thread scheduling). We record the TSC counter value to
ensure that the order of events recorded == order in which they occur.