Read MTC/TSC IntelPT packets when the guest is in non-root & root mode.
This will ensure that the TSC values computed will be correct and reflect
the correct state of the guest, with regards to time. We had to change the linux
kernel so now all of the Intel PT packets are dumped, both when the guest is in root
vs non-root mode.